# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "CareMesh AI (portfolio simulation)", "description": "Portfolio project simulating an AI native youth mental health platform. Not a real healthcare product. Not HIPAA compliant. Not for real patients.", "version": "0.1.0"}, "paths": {"/healthz": {"get": {"tags": ["health"], "summary": "Healthz", "operationId": "healthz_healthz_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Healthz Healthz Get"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenPairResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/refresh": {"post": {"tags": ["auth"], "summary": "Refresh", "operationId": "refresh_api_v1_auth_refresh_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefreshRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenPairResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["auth"], "summary": "Me", "operationId": "me_api_v1_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MeResponse"}}}}}}}, "/api/v1/conversations": {"post": {"tags": ["conversations"], "summary": "Create Conversation", "operationId": "create_conversation_api_v1_conversations_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationCreateRequest"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["conversations"], "summary": "List Conversations", "operationId": "list_conversations_api_v1_conversations_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ConversationResponse"}, "title": "Response List Conversations Api V1 Conversations Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/conversations/{conversation_id}": {"get": {"tags": ["conversations"], "summary": "Get Conversation", "operationId": "get_conversation_api_v1_conversations__conversation_id__get", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Conversation Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/conversations/{conversation_id}/messages": {"get": {"tags": ["conversations"], "summary": "List Messages", "operationId": "list_messages_api_v1_conversations__conversation_id__messages_get", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Conversation Id"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 50, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/MessageResponse"}, "title": "Response List Messages Api V1 Conversations  Conversation Id  Messages Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["conversations"], "summary": "Post Message", "operationId": "post_message_api_v1_conversations__conversation_id__messages_post", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Conversation Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageCreateRequest"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/reviews": {"get": {"tags": ["reviews"], "summary": "List Pending Reviews", "operationId": "list_pending_reviews_api_v1_reviews_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ReviewQueueItemResponse"}, "type": "array", "title": "Response List Pending Reviews Api V1 Reviews Get"}}}}}}}, "/api/v1/reviews/{workflow_id}": {"post": {"tags": ["reviews"], "summary": "Decide Review", "operationId": "decide_review_api_v1_reviews__workflow_id__post", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewDecisionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewDecisionResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/workflows": {"get": {"tags": ["ops"], "summary": "List Workflows", "operationId": "list_workflows_api_v1_ops_workflows_get", "parameters": [{"name": "state", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "State"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 200, "minimum": 1, "default": 50, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/OpsWorkflowResponse"}, "title": "Response List Workflows Api V1 Ops Workflows Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/workflows/{workflow_id}": {"get": {"tags": ["ops"], "summary": "Workflow Detail", "operationId": "workflow_detail_api_v1_ops_workflows__workflow_id__get", "parameters": [{"name": "workflow_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Workflow Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OpsWorkflowDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/ai-requests": {"get": {"tags": ["ops"], "summary": "List Ai Requests", "operationId": "list_ai_requests_api_v1_ops_ai_requests_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 200, "minimum": 1, "default": 50, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/OpsAIRequestResponse"}, "title": "Response List Ai Requests Api V1 Ops Ai Requests Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/ai-requests/{request_id}": {"get": {"tags": ["ops"], "summary": "Ai Request Detail", "operationId": "ai_request_detail_api_v1_ops_ai_requests__request_id__get", "parameters": [{"name": "request_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Request Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OpsAIRequestDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/events": {"get": {"tags": ["ops"], "summary": "List Events", "operationId": "list_events_api_v1_ops_events_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 200, "minimum": 1, "default": 50, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/OpsEventResponse"}, "title": "Response List Events Api V1 Ops Events Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/events/{event_id}/republish": {"post": {"tags": ["ops"], "summary": "Republish Event", "operationId": "republish_event_api_v1_ops_events__event_id__republish_post", "parameters": [{"name": "event_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Event Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ops/dlq": {"get": {"tags": ["ops"], "summary": "View Dlq", "operationId": "view_dlq_api_v1_ops_dlq_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OpsDlqResponse"}}}}}}}, "/api/v1/knowledge/documents": {"get": {"tags": ["knowledge"], "summary": "List Documents", "operationId": "list_documents_api_v1_knowledge_documents_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/KnowledgeDocumentResponse"}, "type": "array", "title": "Response List Documents Api V1 Knowledge Documents Get"}}}}}}, "post": {"tags": ["knowledge"], "summary": "Ingest Document", "operationId": "ingest_document_api_v1_knowledge_documents_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DocumentIngestRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DocumentIngestResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/knowledge/ask": {"post": {"tags": ["knowledge"], "summary": "Ask", "operationId": "ask_api_v1_knowledge_ask_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AskRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AskResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/school/roster": {"get": {"tags": ["school"], "summary": "Roster", "operationId": "roster_api_v1_school_roster_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/RosterEntryResponse"}, "type": "array", "title": "Response Roster Api V1 School Roster Get"}}}}}}}, "/api/v1/school/referrals": {"get": {"tags": ["school"], "summary": "My Referrals", "operationId": "my_referrals_api_v1_school_referrals_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ReferralResponse"}, "type": "array", "title": "Response My Referrals Api V1 School Referrals Get"}}}}}}, "post": {"tags": ["school"], "summary": "Submit Referral", "operationId": "submit_referral_api_v1_school_referrals_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReferralSubmitRequest"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReferralResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/guardian/overview": {"get": {"tags": ["guardian and referrals"], "summary": "Guardian Overview", "operationId": "guardian_overview_api_v1_guardian_overview_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GuardianOverviewResponse"}}}}}}}, "/api/v1/guardian/updates": {"post": {"tags": ["guardian and referrals"], "summary": "Share Guardian Update", "operationId": "share_guardian_update_api_v1_guardian_updates_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GuardianUpdateRequest"}}}, "required": true}, "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/my-patients": {"get": {"tags": ["guardian and referrals"], "summary": "My Patients", "operationId": "my_patients_api_v1_my_patients_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Response My Patients Api V1 My Patients Get"}}}}}}}, "/api/v1/referrals": {"get": {"tags": ["guardian and referrals"], "summary": "Pending Referrals", "operationId": "pending_referrals_api_v1_referrals_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ReferralResponse"}, "type": "array", "title": "Response Pending Referrals Api V1 Referrals Get"}}}}}}}, "/api/v1/referrals/{referral_id}/decision": {"post": {"tags": ["guardian and referrals"], "summary": "Decide Referral", "operationId": "decide_referral_api_v1_referrals__referral_id__decision_post", "parameters": [{"name": "referral_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Referral Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReferralDecideRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Decide Referral Api V1 Referrals  Referral Id  Decision Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AskRequest": {"properties": {"question": {"type": "string", "maxLength": 500, "minLength": 3, "title": "Question"}}, "type": "object", "required": ["question"], "title": "AskRequest"}, "AskResponse": {"properties": {"answer": {"type": "string", "title": "Answer"}, "grounded": {"type": "boolean", "title": "Grounded"}, "simulated": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Simulated"}, "model": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Model"}, "citations": {"items": {"$ref": "#/components/schemas/CitationResponse"}, "type": "array", "title": "Citations"}}, "type": "object", "required": ["answer", "grounded", "simulated", "model", "citations"], "title": "AskResponse"}, "CitationResponse": {"properties": {"chunk_id": {"type": "string", "format": "uuid", "title": "Chunk Id"}, "document_title": {"type": "string", "title": "Document Title"}, "document_version": {"type": "integer", "title": "Document Version"}, "snippet": {"type": "string", "title": "Snippet"}, "score": {"type": "number", "title": "Score"}, "used": {"type": "boolean", "title": "Used"}}, "type": "object", "required": ["chunk_id", "document_title", "document_version", "snippet", "score", "used"], "title": "CitationResponse"}, "ConversationCreateRequest": {"properties": {"title": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Title"}}, "type": "object", "required": ["title"], "title": "ConversationCreateRequest"}, "ConversationResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "patient_id": {"type": "string", "format": "uuid", "title": "Patient Id"}, "title": {"type": "string", "title": "Title"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "patient_id", "title", "created_at"], "title": "ConversationResponse"}, "DocumentIngestRequest": {"properties": {"title": {"type": "string", "maxLength": 300, "minLength": 1, "title": "Title"}, "source_name": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Source Name"}, "content": {"type": "string", "maxLength": 100000, "minLength": 1, "title": "Content"}}, "type": "object", "required": ["title", "source_name", "content"], "title": "DocumentIngestRequest"}, "DocumentIngestResponse": {"properties": {"document": {"$ref": "#/components/schemas/KnowledgeDocumentResponse"}, "chunk_count": {"type": "integer", "title": "Chunk Count"}, "unchanged": {"type": "boolean", "title": "Unchanged"}}, "type": "object", "required": ["document", "chunk_count", "unchanged"], "title": "DocumentIngestResponse"}, "GuardianOverviewResponse": {"properties": {"students": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Students"}, "updates": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Updates"}, "notifications": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Notifications"}}, "type": "object", "required": ["students", "updates", "notifications"], "title": "GuardianOverviewResponse"}, "GuardianUpdateRequest": {"properties": {"patient_id": {"type": "string", "format": "uuid", "title": "Patient Id"}, "content": {"type": "string", "maxLength": 4000, "minLength": 5, "title": "Content"}}, "type": "object", "required": ["patient_id", "content"], "title": "GuardianUpdateRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "KnowledgeDocumentResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "source_name": {"type": "string", "title": "Source Name"}, "version": {"type": "integer", "title": "Version"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "title", "source_name", "version", "created_at"], "title": "KnowledgeDocumentResponse"}, "LoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "MeResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "role": {"$ref": "#/components/schemas/Role"}, "display_name": {"type": "string", "title": "Display Name"}, "organization_id": {"type": "string", "format": "uuid", "title": "Organization Id"}}, "type": "object", "required": ["id", "email", "role", "display_name", "organization_id"], "title": "MeResponse"}, "MessageCreateRequest": {"properties": {"content": {"type": "string", "maxLength": 8000, "minLength": 1, "title": "Content"}}, "type": "object", "required": ["content"], "title": "MessageCreateRequest"}, "MessageResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "conversation_id": {"type": "string", "format": "uuid", "title": "Conversation Id"}, "sender_type": {"$ref": "#/components/schemas/SenderType"}, "sender_id": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Sender Id"}, "content": {"type": "string", "title": "Content"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "simulated": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Simulated"}}, "type": "object", "required": ["id", "conversation_id", "sender_type", "sender_id", "content", "created_at"], "title": "MessageResponse"}, "OpsAIRequestDetailResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "provider": {"type": "string", "title": "Provider"}, "model": {"type": "string", "title": "Model"}, "prompt_name": {"type": "string", "title": "Prompt Name"}, "prompt_version": {"type": "integer", "title": "Prompt Version"}, "status": {"type": "string", "title": "Status"}, "simulated": {"type": "boolean", "title": "Simulated"}, "validation_ok": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Validation Ok"}, "latency_ms": {"type": "number", "title": "Latency Ms"}, "input_tokens": {"type": "integer", "title": "Input Tokens"}, "output_tokens": {"type": "integer", "title": "Output Tokens"}, "cost_usd": {"type": "number", "title": "Cost Usd"}, "correlation_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlation Id"}, "error_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error Type"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "request_messages": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Request Messages"}, "response_text": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Response Text"}}, "type": "object", "required": ["id", "provider", "model", "prompt_name", "prompt_version", "status", "simulated", "validation_ok", "latency_ms", "input_tokens", "output_tokens", "cost_usd", "correlation_id", "error_type", "created_at", "request_messages", "response_text"], "title": "OpsAIRequestDetailResponse"}, "OpsAIRequestResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "provider": {"type": "string", "title": "Provider"}, "model": {"type": "string", "title": "Model"}, "prompt_name": {"type": "string", "title": "Prompt Name"}, "prompt_version": {"type": "integer", "title": "Prompt Version"}, "status": {"type": "string", "title": "Status"}, "simulated": {"type": "boolean", "title": "Simulated"}, "validation_ok": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Validation Ok"}, "latency_ms": {"type": "number", "title": "Latency Ms"}, "input_tokens": {"type": "integer", "title": "Input Tokens"}, "output_tokens": {"type": "integer", "title": "Output Tokens"}, "cost_usd": {"type": "number", "title": "Cost Usd"}, "correlation_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlation Id"}, "error_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error Type"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "provider", "model", "prompt_name", "prompt_version", "status", "simulated", "validation_ok", "latency_ms", "input_tokens", "output_tokens", "cost_usd", "correlation_id", "error_type", "created_at"], "title": "OpsAIRequestResponse"}, "OpsDlqResponse": {"properties": {"topic": {"type": "string", "title": "Topic"}, "records": {"items": {"type": "string"}, "type": "array", "title": "Records"}}, "type": "object", "required": ["topic", "records"], "title": "OpsDlqResponse"}, "OpsEventResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "event_type": {"type": "string", "title": "Event Type"}, "schema_version": {"type": "integer", "title": "Schema Version"}, "occurred_at": {"type": "string", "format": "date-time", "title": "Occurred At"}, "published_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Published At"}, "correlation_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlation Id"}}, "type": "object", "required": ["id", "event_type", "schema_version", "occurred_at", "published_at", "correlation_id"], "title": "OpsEventResponse"}, "OpsTransitionResponse": {"properties": {"from_state": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From State"}, "to_state": {"type": "string", "title": "To State"}, "actor": {"type": "string", "title": "Actor"}, "reason": {"type": "string", "title": "Reason"}, "occurred_at": {"type": "string", "format": "date-time", "title": "Occurred At"}}, "type": "object", "required": ["from_state", "to_state", "actor", "reason", "occurred_at"], "title": "OpsTransitionResponse"}, "OpsWorkflowDetailResponse": {"properties": {"workflow": {"$ref": "#/components/schemas/OpsWorkflowResponse"}, "transitions": {"items": {"$ref": "#/components/schemas/OpsTransitionResponse"}, "type": "array", "title": "Transitions"}}, "type": "object", "required": ["workflow", "transitions"], "title": "OpsWorkflowDetailResponse"}, "OpsWorkflowResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "workflow_type": {"type": "string", "title": "Workflow Type"}, "state": {"type": "string", "title": "State"}, "subject_id": {"type": "string", "format": "uuid", "title": "Subject Id"}, "correlation_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Correlation Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "workflow_type", "state", "subject_id", "correlation_id", "created_at", "updated_at"], "title": "OpsWorkflowResponse"}, "ReferralDecideRequest": {"properties": {"accept": {"type": "boolean", "title": "Accept"}}, "type": "object", "required": ["accept"], "title": "ReferralDecideRequest"}, "ReferralResponse": {"properties": {"referral_id": {"type": "string", "format": "uuid", "title": "Referral Id"}, "workflow_id": {"type": "string", "format": "uuid", "title": "Workflow Id"}, "patient_id": {"type": "string", "format": "uuid", "title": "Patient Id"}, "patient_name": {"type": "string", "title": "Patient Name"}, "state": {"type": "string", "title": "State"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "concern": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Concern"}}, "type": "object", "required": ["referral_id", "workflow_id", "patient_id", "patient_name", "state", "created_at", "concern"], "title": "ReferralResponse"}, "ReferralSubmitRequest": {"properties": {"patient_id": {"type": "string", "format": "uuid", "title": "Patient Id"}, "concern": {"type": "string", "maxLength": 4000, "minLength": 10, "title": "Concern"}, "consent_confirmed": {"type": "boolean", "title": "Consent Confirmed"}}, "type": "object", "required": ["patient_id", "concern", "consent_confirmed"], "title": "ReferralSubmitRequest"}, "RefreshRequest": {"properties": {"refresh_token": {"type": "string", "title": "Refresh Token"}}, "type": "object", "required": ["refresh_token"], "title": "RefreshRequest"}, "ReviewDecision": {"type": "string", "enum": ["accept", "edit", "reject"], "title": "ReviewDecision"}, "ReviewDecisionRequest": {"properties": {"decision": {"$ref": "#/components/schemas/ReviewDecision"}, "severity_override": {"anyOf": [{"type": "integer", "maximum": 3.0, "minimum": 0.0}, {"type": "null"}], "title": "Severity Override"}, "note": {"type": "string", "maxLength": 2000, "title": "Note", "default": ""}}, "type": "object", "required": ["decision"], "title": "ReviewDecisionRequest"}, "ReviewDecisionResponse": {"properties": {"workflow_id": {"type": "string", "format": "uuid", "title": "Workflow Id"}, "state": {"type": "string", "title": "State"}, "decision": {"$ref": "#/components/schemas/ReviewDecision"}}, "type": "object", "required": ["workflow_id", "state", "decision"], "title": "ReviewDecisionResponse"}, "ReviewQueueItemResponse": {"properties": {"workflow_id": {"type": "string", "format": "uuid", "title": "Workflow Id"}, "risk_signal_id": {"type": "string", "format": "uuid", "title": "Risk Signal Id"}, "patient_name": {"type": "string", "title": "Patient Name"}, "message_content": {"type": "string", "title": "Message Content"}, "category": {"$ref": "#/components/schemas/RiskCategory"}, "severity": {"type": "integer", "title": "Severity"}, "confidence": {"type": "number", "title": "Confidence"}, "evidence": {"type": "string", "title": "Evidence"}, "model": {"type": "string", "title": "Model"}, "prompt_version": {"type": "integer", "title": "Prompt Version"}, "simulated": {"type": "boolean", "title": "Simulated"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["workflow_id", "risk_signal_id", "patient_name", "message_content", "category", "severity", "confidence", "evidence", "model", "prompt_version", "simulated", "created_at"], "title": "ReviewQueueItemResponse"}, "RiskCategory": {"type": "string", "enum": ["none", "low_mood", "anxiety", "self_harm", "crisis"], "title": "RiskCategory"}, "Role": {"type": "string", "enum": ["patient", "guardian", "therapist", "school_staff", "payer_staff", "ops_admin"], "title": "Role"}, "RosterEntryResponse": {"properties": {"patient_id": {"type": "string", "format": "uuid", "title": "Patient Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["patient_id", "name"], "title": "RosterEntryResponse"}, "SenderType": {"type": "string", "enum": ["patient", "dira", "clinician", "system"], "title": "SenderType"}, "TokenPairResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "refresh_token": {"type": "string", "title": "Refresh Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}, "expires_in": {"type": "integer", "title": "Expires In"}}, "type": "object", "required": ["access_token", "refresh_token", "expires_in"], "title": "TokenPairResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}